NEWS	R Documentation

News for package ‘pomp’

Changes in pomp version 1.6.1

Major revisions to the package help pages and manual. The separate documentation on C snippets and process-model plugins has been merged into the main pomp help page, which now has comprehensive instructions on building a pomp object.

Update to the package help page.

Changes in pomp version 1.5.10

The matrix and numeric vector options for the data argument to pomp are now deprecated and will be removed in a future release. In calls to pomp, data should be either a data frame or a pomp-class object.

Many more errors are now trapped and reported with more informative messages.

Changes in pomp version 1.5.9

Redundant bootstrap argument to nlf has been removed.

More improvements to the testing suite.

Changes in pomp version 1.5.8

Major improvements to error and warning handling.

Fix bugs that might arise in rare circumstances in simulate and eulermultinom functions.

More improvements to the testing suite.

Changes in pomp version 1.5.7

Many improvements to the test suite.

Fixed a bug in pompExample.

Changes in pomp version 1.5.6

Major overhaul of error handling system and error messages.

There is a new plugin for rprocess called kleap.sim. This implements the so-called “K-leap method” of Cai & Xu (2007).

Changes in pomp version 1.5.5

Many improvements to the test suite.

Changes in pomp version 1.5.4

More improvements to the test suite.

Fixed bugs found as a result of improved testing.

In mif2, it is now required that Nmif>0.

Changes in pomp version 1.5.3

More improvements to the test suite.

Fixed bugs found as a result of improved testing.

Fixed bug in specification of skeleton in pompExamples.

Changes in pomp version 1.5.2

Improved the test suite to obtain better code coverage.

pomp is no longer an S4 method. Its new status as an ordinary function abbreviates and consolidates the code and simplifies the documentation.

Some minor improvements to error messages.

A bug restricting access in C code to userdata has been fixed.

Fixed bug in pompExample(...,show=TRUE).

Changes in pomp version 1.5.1

New methods for the Ensemble Kalman Filter (enkf) and Ensemble Adjustment Kalman Filter (eakf) have been added to the package.

Changes in pomp version 1.4.9

A better interface for specifying a model's deterministic skeleton is provided. One specifies skeleton=map(f,delta.t) for a discrete-time skeleton (a map) and skeleton=vectorfield(f) for a continuous-time skeleton (a vectorfield). The old arguments skeleton.type and skelmap.delta.t are deprecated and will be removed in a future release.

The pfilter help page has been improved. Specifically, the discussion of filtering failures is better.

Miscellaneous improvements to the help pages.

Changes in pomp version 1.4.8

mif2 no longer computes filter means.

Changes in pomp version 1.4.7

The new argument show of pompExamples allows one to display the example code instead of executing it.

The minimum version of R supported is now 3.1.2.

Changes in pomp version 1.4.6

init.state now has the optional argument nsim. Using this, one can request multiple initial state vectors per parameter vector.

pfilter now uses less memory when it is asked to run at a single point in parameter space.

Changes in pomp version 1.4.5

A bug associated with the rw.sd argument to mif2 on Windows platforms has been fixed.

Changes in pomp version 1.4.4

The method="mif2" option to mif has been removed. Use mif2 instead.

Changes in pomp version 1.4.3

The particles method (rarely if ever used), has been removed to streamline the mif codes.

Changes in pomp version 1.4.2

The method="mif2" option to mif is now deprecated. It will be removed in a future release. Use mif2 instead.

Changes in pomp version 1.4.1

Following the publication of the Journal of Statistical Software announcement paper, the ‘CITATION’ has been modified to include the appropriate citation. Please use citation("pomp") to learn about how to cite pomp correctly.

Changes in pomp version 1.3.4

conv.rec on the results of a mif2 computation now play nicely with reshape2::melt.

Changes in pomp version 1.3.3

It is now possible to direct pomp to place Csnippet codes in a specified location. This is accomplished via the new cdir and cfile arguments to pomp.

Changes in pomp version 1.3.2

When several concurrent R processes attempt to simultaneously compile Csnipet code from copies of a pomp object on the same machine, a race condition exists. This release fixes this by ensuring that Csnippet codes are written and compiled in a process-specific directory.

Changes in pomp version 1.3.1

This release removes several long-deprecated features.

The seed argument to pfilter has been removed. Use freeze to obtain similar functionality.

The redundant pars argument to mif has been removed. Names of parameters to be estimated are taken from the nonzero entries of rw.sd. Initial-value parameters are designated using the ivps argument.

The old parameter.transform and parameter.inv.transform arguments to pomp have been removed. They were superseded by the arguments fromEstimationScale and toEstimationScale in version 0.65-1.

Changes in pomp version 1.2.2

Some internal changes to mif2 by C. Breto, in support of panel methods under development.

Changes in pomp version 1.2.1

Bug fix: for POMPs with a scalar state variable, the filter.mean, pred.mean, and pred.var methods dropped the singleton dimension of the respective arrays, in conflict with the assumptions of certain plot methods. This resolves issue #3.

Changes in pomp version 1.1.4

Documentation improvements. Specifically, the details of how step-sizes are chosen in the euler.sim, discrete.time.sim, and onestep.sim plugins has been spelled out explicitly. Also, incompleteness in the sannbox documentation has been corrected.

Changes in pomp version 1.1.3

Provide filter.traj method for pmcmcList objects.

Changes in pomp version 1.1.2

Fixed a bug in mif and mif2 when cooling.type="hyperbolic" and cooling.fraction.50=1.