basic_probes {pomp}	R Documentation

Useful probes for partially-observed Markov processes

Description

Several simple and configurable probes are provided with in the package. These can be used directly and as templates for custom probes.

Usage

probe_mean(var, trim = 0, transform = identity, na.rm = TRUE)

probe_median(var, na.rm = TRUE)

probe_var(var, transform = identity, na.rm = TRUE)

probe_sd(var, transform = identity, na.rm = TRUE)

probe_period(var, kernel.width, transform = identity)

probe_quantile(var, probs, ...)

probe_acf(
  var,
  lags,
  type = c("covariance", "correlation"),
  transform = identity
)

probe_ccf(
  vars,
  lags,
  type = c("covariance", "correlation"),
  transform = identity
)

probe_marginal(var, ref, order = 3, diff = 1, transform = identity)

probe_nlar(var, lags, powers, transform = identity)

Arguments

var, vars
character; the name(s) of the observed variable(s).

trim
the fraction of observations to be trimmed (see mean).

transform
transformation to be applied to the data before the probe is computed.

na.rm
if TRUE, remove all NA observations prior to computing the probe.

kernel.width
width of modified Daniell smoothing kernel to be used in power-spectrum computation: see kernel.

probs
the quantile or quantiles to compute: see quantile.

...
additional arguments passed to the underlying algorithms.

lags
In probe_ccf, a vector of lags between time series. Positive lags correspond to x advanced relative to y; negative lags, to the reverse.

In probe_nlar, a vector of lags present in the nonlinear autoregressive model that will be fit to the actual and simulated data. See Details, below, for a precise description.

type
Compute autocorrelation or autocovariance?

ref
empirical reference distribution. Simulated data will be regressed against the values of ref, sorted and, optionally, differenced. The resulting regression coefficients capture information about the shape of the marginal distribution. A good choice for ref is the data itself.

order
order of polynomial regression.

diff
order of differencing to perform.

powers
the powers of each term (corresponding to lags) in the the nonlinear autoregressive model that will be fit to the actual and simulated data. See Details, below, for a precise description.

Value

A call to any one of these functions returns a probe function, suitable for use in probe or probe_objfun. That is, the function returned by each of these takes a data array (such as comes from a call to obs) as input and returns a single numerical value.

Author(s)

Daniel C. Reuman, Aaron A. King

References

B.E. Kendall, C.J. Briggs, W.W. Murdoch, P. Turchin, S.P. Ellner, E. McCauley, R.M. Nisbet, and S.N. Wood. Why do populations cycle? A synthesis of statistical and mechanistic modeling approaches. Ecology 80, 1789–1805, 1999.

S. N. Wood Statistical inference for noisy nonlinear ecological dynamic systems. Nature 466, 1102–1104, 2010.

See Also

More on methods based on summary statistics: abc(), nlf, probe(), probe_match, spect(), spect_match

[Package pomp version 5.7.1.0 Index]