cond.logLik {pomp}	R Documentation

Conditional log likelihood

Description

The estimated conditional log likelihood from a fitted model.

Usage

## S4 method for signature 'kalmand_pomp'
cond.logLik(object, ...)

## S4 method for signature 'pfilterd_pomp'
cond.logLik(object, ...)

## S4 method for signature 'wpfilterd_pomp'
cond.logLik(object, ...)

## S4 method for signature 'bsmcd_pomp'
cond.logLik(object, ...)

Arguments

object
result of a filtering computation

...
ignored

Details

The conditional likelihood is defined to be the value of the density of

Yk | Y1,…,Y(k-1)

evaluated at Yk = yk*. Here, Yk is the observable process, and yk* the data, at time t_k.

Thus the conditional log likelihood at time t_k is

ell_k(theta)=log f[Yk = yk* | Y1=y1*, …, Y(k-1)=y(k-1)*],

where f is the probability density above.

Value

The numerical value of the conditional log likelihood. Note that some methods compute not the log likelihood itself but instead a related quantity. To keep the code simple, the cond.logLik function is nevertheless used to extract this quantity.

When object is of class ‘bsmcd_pomp’ (i.e., the result of a bsmc2 computation), cond.logLik returns the conditional log “evidence” (see bsmc2).

See Also

More on sequential Monte Carlo methods: bsmc2(), eff.sample.size(), filter.mean(), filter.traj(), kalman, mif2(), pfilter(), pmcmc(), pred.mean(), pred.var(), saved.states(), wpfilter()

Other extraction methods: coef(), covmat(), eff.sample.size(), filter.mean(), filter.traj(), forecast(), logLik, obs(), pred.mean(), pred.var(), saved.states(), spy(), states(), summary(), timezero(), time(), traces()

[Package pomp version 4.1.0.0 Index]