dacca {pomp}	R Documentation

Model of cholera transmission for historic Bengal.

Description

dacca constructs a ‘pomp’ object containing census and cholera mortality data from the Dacca district of the former British province of Bengal over the years 1891 to 1940 together with a stochastic differential equation transmission model. The model is that of King et al. (2008). The parameters are the MLE for the SIRS model with seasonal reservoir.

Usage

dacca(
  gamma = 20.8,
  eps = 19.1,
  rho = 0,
  delta = 0.02,
  deltaI = 0.06,
  clin = 1,
  alpha = 1,
  beta_trend = -0.00498,
  logbeta = c(0.747, 6.38, -3.44, 4.23, 3.33, 4.55),
  logomega = log(c(0.184, 0.0786, 0.0584, 0.00917, 0.000208, 0.0124)),
  sd_beta = 3.13,
  tau = 0.23,
  S_0 = 0.621,
  I_0 = 0.378,
  Y_0 = 0,
  R1_0 = 0.000843,
  R2_0 = 0.000972,
  R3_0 = 1.16e-07
)

Arguments

gamma
recovery rate

eps
rate of waning of immunity for severe infections

rho
rate of waning of immunity for inapparent infections

delta
baseline mortality rate

deltaI
cholera mortality rate

clin
fraction of infections that lead to severe infection

alpha
transmission function exponent

beta_trend
slope of secular trend in transmission

logbeta
seasonal transmission rates

logomega
seasonal environmental reservoir parameters

sd_beta
environmental noise intensity

tau
measurement error s.d.

S_0
initial susceptible fraction

I_0
initial fraction of population infected

Y_0
initial fraction of the population in the Y class

R1_0, R2_0, R3_0
initial fractions in the respective R classes

Details

Data are provided courtesy of Dr. Menno J. Bouma, London School of Tropical Medicine and Hygiene.

Value

dacca returns a ‘pomp’ object containing the model, data, and MLE parameters, as estimated by King et al. (2008).

References

A.A. King, E.L. Ionides, M. Pascual, and M.J. Bouma. Inapparent infections and cholera dynamics. Nature 454, 877-880, 2008

See Also

More examples provided with pomp: SIR models, blowflies, childhood disease data, ebola, gompertz(), ou2(), pomp examples, ricker(), rw2(), verhulst()

More data sets provided with pomp: blowflies, bsflu, childhood disease data, ebola, parus

Examples

## Not run: 
  po <- dacca()
  plot(po)
  ## MLE:
  coef(po)
  plot(simulate(po))

## End(Not run)

[Package pomp version 4.5.0.0 Index]