cond.loglik {pomp}	R Documentation

Deprecated functions.

Description

These functions are obsolete and will be removed in a future release.

Obsolete: replaced by profile_design.

Obsolete: replaced by runif_design.

Obsolete: replaced by slice_design.

Obsolete: replaced by sobol_design.

Usage

cond.loglik(...)

profileDesign(
  ...,
  lower,
  upper,
  nprof,
  type = c("sobol", "runif"),
  stringsAsFactors = getOption("stringsAsFactors", FALSE)
)

runifDesign(lower = numeric(0), upper = numeric(0), nseq)

sliceDesign(center, ...)

sobolDesign(lower = numeric(0), upper = numeric(0), nseq)

Arguments

...
In profileDesign, additional arguments specify the parameters over which to profile and the values of these parameters. In sliceDesign, additional numeric vector arguments specify the locations of points along the slices.

lower, upper
named numeric vectors giving the lower and upper bounds of the ranges, respectively.

nprof
The number of points per profile point.

type
the type of design to use. type="sobol" uses sobolDesign; type="runif" uses runifDesign.

stringsAsFactors
should character vectors be converted to factors?

nseq
Total number of points requested.

center
center is a named numeric vector specifying the point through which the slice(s) is (are) to be taken.

Details

The Sobol' sequence generation is performed using codes from the NLopt library by S. Johnson.

Value

profileDesign returns a data frame with nprof points per profile point.

runifDesign returns a data frame with nseq rows and one column for each variable named in lower and upper.

sliceDesign returns a data frame with one row per point. The ‘slice’ variable indicates which slice the point belongs to.

sobolDesign returns a data frame with nseq rows and one column for each variable named in lower and upper.

Author(s)

Aaron A. King

[Package pomp version 3.3.0.0 Index]