filter.traj {pomp}	R Documentation

Filtering trajectories

Description

Drawing from the smoothing distribution

Usage

## S4 method for signature 'pfilterd_pomp'
filter.traj(object, vars, ...)

## S4 method for signature 'pfilterList'
filter.traj(object, vars, ...)

## S4 method for signature 'pmcmcd_pomp'
filter.traj(object, vars, ...)

## S4 method for signature 'pmcmcList'
filter.traj(object, vars, ...)

Arguments

object
result of a filtering computation

vars
optional character; names of variables

...
ignored

Details

The smoothing distribution is the distribution of

X(t_k) | Y(t_1)=y^*_1, \dots, Y(t_n)=y^*_n,

where X(t_k) is the latent state process and Y(t_k) is the observable process at time t_k, and n is the number of observations.

To draw samples from this distribution, one can run a number of independent particle filter (pfilter) operations, sampling the full trajectory of one randomly-drawn particle from each one. One should view these as weighted samples from the smoothing distribution, where the weights are the likelihoods returned by each of the pfilter computations.

One accomplishes this by setting filter.traj = TRUE in each pfilter computation and extracting the trajectory using the filter.traj command.

In particle MCMC (pmcmc), the tracking of an individual trajectory is performed automatically.

See Also

More on sequential Monte Carlo methods: bsmc2(), cond.logLik(), eff.sample.size(), filter.mean(), kalman, mif2(), pfilter(), pmcmc(), pred.mean(), pred.var(), saved.states(), wpfilter()

Other extraction methods: coef(), cond.logLik(), covmat(), eff.sample.size(), filter.mean(), forecast(), logLik, obs(), pred.mean(), pred.var(), saved.states(), spy(), states(), summary(), timezero(), time(), traces()

[Package pomp version 4.2.3.0 Index]