lookup {pomp}	R Documentation

Lookup table

Description

Interpolate values from a lookup table

Usage

lookup(table, t)

Arguments

table
a ‘covartable’ object created by a call to covariate_table

t
numeric vector; times at which interpolated values of the covariates in table are required.

Details

A warning will be generated if extrapolation is performed.

Value

A numeric vector or matrix of the interpolated values.

See Also

More on interpolation: bsplines, covariates

[Package pomp version 4.2.2.0 Index]