melt {pomp}	R Documentation

Melt

Description

Convert arrays, lists, and other objects to data frames.

Usage

## S4 method for signature 'ANY'
melt(data, ...)

## S4 method for signature 'array'
melt(data, ...)

## S4 method for signature 'list'
melt(data, ..., level = 1)

Arguments

data
object to convert

...
ignored

level
integer; level of recursion

Details

melt converts its first argument to a data frame. It is a simplified version of the melt command provided by the no-longer maintained reshape2 package.

An array can be melted into a data frame. In this case, the data frame will have one row per entry in the array.

A list can be melted into a data frame. This operation is recursive. A variable will be appended to distinguish the separate list entries.

[Package pomp version 5.8.2.1 Index]