mif2 {pomp}	R Documentation

Iterated filtering: maximum likelihood by iterated, perturbed Bayes maps

Description

An iterated filtering algorithm for estimating the parameters of a partially-observed Markov process. Running mif2 causes the algorithm to perform a specified number of particle-filter iterations. At each iteration, the particle filter is performed on a perturbed version of the model, in which the parameters to be estimated are subjected to random perturbations at each observation. This extra variability effectively smooths the likelihood surface and combats particle depletion by introducing diversity into particle population. As the iterations progress, the magnitude of the perturbations is diminished according to a user-specified cooling schedule. The algorithm is presented and justified in Ionides et al. (2015).

Usage

## S4 method for signature 'data.frame'
mif2(
  data,
  ...,
  Nmif = 1,
  rw.sd,
  cooling.type = c("geometric", "hyperbolic"),
  cooling.fraction.50,
  Np,
  params,
  rinit,
  rprocess,
  dmeasure,
  partrans,
  verbose = getOption("verbose", FALSE)
)

## S4 method for signature 'pomp'
mif2(
  data,
  ...,
  Nmif = 1,
  rw.sd,
  cooling.type = c("geometric", "hyperbolic"),
  cooling.fraction.50,
  Np,
  verbose = getOption("verbose", FALSE)
)

## S4 method for signature 'pfilterd_pomp'
mif2(data, ..., Nmif = 1, Np, verbose = getOption("verbose", FALSE))

## S4 method for signature 'mif2d_pomp'
mif2(
  data,
  ...,
  Nmif,
  rw.sd,
  cooling.type,
  cooling.fraction.50,
  verbose = getOption("verbose", FALSE)
)

Arguments

data
either a data frame holding the time series data, or an object of class ‘pomp’, i.e., the output of another pomp calculation. Internally, data will be coerced to an array with storage-mode double.

...
additional arguments are passed to pomp. This allows one to set, unset, or modify basic model components within a call to this function.

Nmif
The number of filtering iterations to perform.

rw.sd
specification of the magnitude of the random-walk perturbations that will be applied to some or all model parameters. Parameters that are to be estimated should have positive perturbations specified here. The specification is given using the rw_sd function, which creates a list of unevaluated expressions. The latter are evaluated in a context where vector of observation times is visible (as ‘time’). The expression ivp(s) can be used in this context as shorthand for

ifelse(time==time[1],s,0).

Likewise, ivp(s,lag) is equivalent to

ifelse(time==time[lag],s,0).

See below for some examples.

The perturbations that are applied are normally distributed with the specified s.d. If parameter transformations have been supplied, then the perturbations are applied on the transformed (estimation) scale.

cooling.type, cooling.fraction.50
specifications for the cooling schedule, i.e., the manner and rate with which the intensity of the parameter perturbations is reduced with successive filtering iterations. cooling.type specifies the nature of the cooling schedule. See below (under “Specifying the perturbations”) for more detail.

Np
the number of particles to use. This may be specified as a single positive integer, in which case the same number of particles will be used at each timestep. Alternatively, if one wishes the number of particles to vary across timesteps, one may specify Np either as a vector of positive integers of length

length(time(object,t0=TRUE))

or as a function taking a positive integer argument. In the latter case, Np(k) must be a single positive integer, representing the number of particles to be used at the k-th timestep: Np(0) is the number of particles to use going from timezero(object) to time(object)[1], Np(1), from timezero(object) to time(object)[1], and so on, while when T=length(time(object)), Np(T) is the number of particles to sample at the end of the time-series.

params
optional; named numeric vector of parameters. This will be coerced internally to storage mode double.

rinit
simulator of the initial-state distribution. This can be furnished either as a C snippet, an R function, or the name of a pre-compiled native routine available in a dynamically loaded library. Setting rinit=NULL sets the initial-state simulator to its default. For more information, see rinit specification.

rprocess
simulator of the latent state process, specified using one of the rprocess plugins. Setting rprocess=NULL removes the latent-state simulator. For more information, see rprocess specification for the documentation on these plugins.

dmeasure
evaluator of the measurement model density, specified either as a C snippet, an R function, or the name of a pre-compiled native routine available in a dynamically loaded library. Setting dmeasure=NULL removes the measurement density evaluator. For more information, see dmeasure specification.

partrans
optional parameter transformations, constructed using parameter_trans.

Many algorithms for parameter estimation search an unconstrained space of parameters. When working with such an algorithm and a model for which the parameters are constrained, it can be useful to transform parameters. One should supply the partrans argument via a call to parameter_trans. For more information, see parameter_trans. Setting partrans=NULL removes the parameter transformations, i.e., sets them to the identity transformation.

verbose
logical; if TRUE, diagnostic messages will be printed to the console.

Value

Upon successful completion, mif2 returns an object of class ‘mif2d_pomp’.

Number of particles

If Np is anything other than a constant, the user must take care that the number of particles requested at the end of the time series matches that requested at the beginning. In particular, if T=length(time(object)), then one should have Np[1]==Np[T+1] when Np is furnished as an integer vector and Np(0)==Np(T) when Np is furnished as a function.

Methods

The following methods are available for such an object:

continue

picks up where mif2 leaves off and performs more filtering iterations.

logLik

returns the so-called mif log likelihood which is the log likelihood of the perturbed model, not of the focal model itself. To obtain the latter, it is advisable to run several pfilter operations on the result of a mif2 computatation.

coef

extracts the point estimate

eff_sample_size

extracts the effective sample size of the final filtering iteration

Various other methods can be applied, including all the methods applicable to a pfilterd_pomp object and all other pomp estimation algorithms and diagnostic methods.

Specifying the perturbations

The rw_sd function simply returns a list containing its arguments as unevaluated expressions. These are then evaluated in a context in which the vector of observation times is defined (as ‘time’). This allows for easy specification of the structure of the perturbations that are to be applied. For example,

    rw_sd(
      a=0.05,
      b=ifelse(time==time[1], 0.2, 0),
      c=ivp(0.2),
      d=ifelse(time==time[13], 0.2, 0),
      e=ivp(0.2, lag=13),
      f=ifelse(time<23, 0.02, 0),
      g=ifelse(time>=23 & time<50, 0.02, 0),
      h=ivp(0.1,lags=3:8)
    )

results in random perturbations of parameter a with s.d. 0.05 at every time step, while parameters b and c both get perturbations of s.d. 0.2 only before the first observation (i.e., at the zero-time). Parameters d and e, by contrast, get perturbations of s.d. 0.2 only before the thirteenth observation. Parameter f gets a random perturbation of size 0.02 before every observation falling before t=23, while g gets perturbed before all observations that fall in the interval 23\le{t}<{50}. Finally, the magnitude of the perturbation of parameter h is applied before the third through eighth observations.

On the m-th IF2 iteration, prior to time-point n, the d-th parameter is given a random increment normally distributed with mean 0 and standard deviation c_{m,n} \sigma_{d,n}, where c is the cooling schedule and \sigma is specified using rw_sd, as described above. Let N be the length of the time series and \alpha=cooling.fraction.50. Then, when cooling.type="geometric", we have

c_{m,n}=\alpha^{\frac{n-1+(m-1)N}{50N}}.

When cooling.type="hyperbolic", we have

c_{m,n}=\frac{s+1}{s+n+(m-1)N},

where s satisfies

\frac{s+1}{s+50N}=\alpha.

Thus, in either case, the perturbations at the end of 50 IF2 iterations are a fraction \alpha smaller than they are at first.

Re-running IF2 iterations

To re-run a sequence of IF2 iterations, one can use the mif2 method on a ‘mif2d_pomp’ object. By default, the same parameters used for the original IF2 run are re-used (except for verbose, the default of which is shown above). If one does specify additional arguments, these will override the defaults.

Note for Windows users

Some Windows users report problems when using C snippets in parallel computations. These appear to arise when the temporary files created during the C snippet compilation process are not handled properly by the operating system. To circumvent this problem, use the cdir and cfile options to cause the C snippets to be written to a file of your choice, thus avoiding the use of temporary files altogether.

Author(s)

Aaron A. King, Edward L. Ionides, Dao Nguyen

References

E.L. Ionides, D. Nguyen, Y. Atchadé, S. Stoev, and A.A. King. Inference for dynamic and latent variable models via iterated, perturbed Bayes maps. Proceedings of the National Academy of Sciences 112, 719–724, 2015. doi:10.1073/pnas.1410597112.

See Also

More on full-information (i.e., likelihood-based) methods: bsmc2(), pfilter(), pmcmc(), wpfilter()

More on sequential Monte Carlo methods: bsmc2(), cond_logLik(), eff_sample_size(), filter_mean(), filter_traj(), kalman, pfilter(), pmcmc(), pred_mean(), pred_var(), saved_states(), wpfilter()

More on pomp estimation algorithms: abc(), bsmc2(), estimation_algorithms, nlf, pmcmc(), pomp-package, probe_match, spect_match

More on maximization-based estimation methods: nlf, probe_match, spect_match, traj_match

[Package pomp version 6.3.0.0 Index]