objfun {pomp}	R Documentation

Objective functions

Description

Methods common to pomp stateful objective functions

Important Note

Since pomp cannot guarantee that the final call an optimizer makes to the function is a call at the optimum, it cannot guarantee that the parameters stored in the function are the optimal ones. Therefore, it is a good idea to evaluate the function on the parameters returned by the optimization routine, which will ensure that these parameters are stored.

[Package pomp version 4.4.3.0 Index]