pStop {pomp}	R Documentation

pStop

Description

Custom error, warning, and message functions.

Usage

pStop(..., who = -1L)

pStop_(...)

pWarn(..., who = -1L)

pWarn_(...)

pMess(..., who = -1L)

pMess_(...)

Arguments

...
message

who
integer or character. If who is an integer, it is passed to sys.call to retrieve the name of the calling function. One can also pass the name of the calling function in who. In either case, the name of the calling function is included in the message.

[Package pomp version 5.2.1.0 Index]