particles-mif {pomp}	R Documentation

Generate particles from the user-specified distribution.

Description

Generate particles from the user-specified distribution. This is part of the low-level interface, used by mif. This help page does not give instruction on how to write a valid particles function: see the documentation for mif instead.

Usage

particles(object, ...)
## S4 method for signature 'mif'
particles(object, Np = 1, center = coef(object), sd = 0, ...)

Arguments

object
the mif object

Np
the number of particles, i.e., number of draws.

center
the central value of the distribution of particles

sd
the width of the distribution

...
additional arguments. At present, these are ignored.

Details

The particles method is used to set up the initial distribution of particles. It is an interface to the user-specifed particles slot in the mif object.

Value

particles returns a list of two matrices. states contains the state-variable portion of the particles; params contains the parameter portion. Each has Np columns.

Author(s)

Aaron A. King kingaa at umich dot edu

See Also

mif, mif-methods, pomp, pomp-class

[Package pomp version 0.62-5 Index]