pomp-fun {pomp}	R Documentation

Definition and methods of the "pomp.fun" class

Description

Definition and methods of the pomp.fun class

Usage

## S4 method for signature 'function'
pomp.fun(f, proto = NULL, ...)
## S4 method for signature 'character'
pomp.fun(f, PACKAGE = NULL, 
      obsnames = character(0), statenames = character(0),
      paramnames = character(0), covarnames = character(0), ...)
## S4 method for signature 'Csnippet'
pomp.fun(f, slotname = NULL, libname = NULL,
      obsnames = character(0), statenames = character(0),
      paramnames = character(0), covarnames = character(0), ...)
## S4 method for signature 'pomp.fun'
pomp.fun(f, ...)
## S4 method for signature 'missing'
pomp.fun(f, ...)
## S4 method for signature 'NULL'
pomp.fun(f, ...)
## S4 method for signature 'pomp.fun'
show(object)
## S4 method for signature 'pomp.fun'
print(x, ...)

Arguments

f
A function or the name of a native routine.

PACKAGE
optional; the name of the dynamically-loadable library in which the native function f can be found.

proto
optional string; a prototype against which f will be checked.

object, x
The pomp.fun object.

Details

The pomp.fun class helps to settle common issues associated with user-defined functions which can be defined either via R code or by a native, compiled routine. It is not exported to userland.

Author(s)

Aaron A. King kingaa at umich dot edu

See Also

pomp

[Package pomp version 0.62-5 Index]