pomp_examples {pomp}	R Documentation

pre-built pomp examples

Description

Examples of pomp objects containing models and data.

Details

pomp includes a number of pre-built examples of pomp objects and data that can be analyzed using pomp methods. These include:

blowflies

Data from Nicholson's experiments with sheep blowfly populations

blowflies1()

A pomp object with some of the blowfly data together with a discrete delay equation model.

blowflies2()

A variant of blowflies1.

bsflu

Data from an outbreak of influenza in a boarding school.

dacca()

Fifty years of census and cholera mortality data, together with a stochastic differential equation transmission model (King et al. 2008).

ebolaModel()

Data from the 2014 West Africa outbreak of Ebola virus disease, together with simple transmission models (King et al. 2015).

gompertz()

The Gompertz population dynamics model, with simulated data.

LondonYorke

Data on incidence of several childhood diseases (London and Yorke 1973)

ewmeas

Measles incidence data from England and Wales

ewcitmeas

Measles incidence data from 7 English cities

ou2()

A 2-D Ornstein-Uhlenbeck process with simulated data

parus

Population censuses of a Parus major population in Wytham Wood, England.

ricker

The Ricker population dynamics model, with simulated data

rw2

A 2-D Brownian motion model, with simulated data.

sir()

A simple continuous-time Markov chain SIR model, coded using Euler-multinomial steps, with simulated data.

sir2()

A simple continuous-time Markov chain SIR model, coded using Gillespie's algorithm, with simulated data.

verhulst()

The Verhulst-Pearl (logistic) model, a continuous-time model of population dynamics, with simulated data

See also the tutorials on the package website for more examples.

References

Anonymous. Influenza in a boarding school. British Medical Journal 1, 587, 1978.

A.A. King, E.L. Ionides, M. Pascual, and M.J. Bouma. Inapparent infections and cholera dynamics. Nature 454, 877-880, 2008

A.A. King, M. Domenech de Cellès, F.M.G. Magpantay, and P. Rohani. Avoidable errors in the modelling of outbreaks of emerging pathogens, with special reference to Ebola. Proceedings of the Royal Society of London, Series B 282, 20150347, 2015.

W. P. London and J. A. Yorke, Recurrent outbreaks of measles, chickenpox and mumps: I. Seasonal variation in contact rates. American Journal of Epidemiology 98, 453–468, 1973.

A.J. Nicholson. The self-adjustment of populations to change. Cold Spring Harbor Symposia on Quantitative Biology 22, 153–173, 1957.

See Also

More examples provided with pomp: blowflies, childhood_disease_data, compartmental_models, dacca(), ebola, gompertz(), ou2(), ricker(), rw2(), verhulst()

[Package pomp version 5.3.1.2 Index]