pred_mean {pomp}	R Documentation

Prediction mean

Description

The mean of the prediction distribution

Usage

## S4 method for signature 'kalmand_pomp'
pred_mean(object, vars, ..., format = c("array", "data.frame"))

## S4 method for signature 'pfilterd_pomp'
pred_mean(object, vars, ..., format = c("array", "data.frame"))

Arguments

object
result of a filtering computation

vars
optional character; names of variables

...
ignored

format
format of the returned object

Details

The prediction distribution is that of

X(t_k) \vert Y(t_1)=y^*_1,\dots,Y(t_{k-1})=y^*_{k-1},

where X(t_k), Y(t_k) are the latent state and observable processes, respectively, and y^*_k is the data, at time t_k.

The prediction mean is therefore the expectation of this distribution

E[X(t_k) \vert Y(t_1)=y^*_1,\dots,Y(t_{k-1})=y^*_{k-1}].

See Also

More on sequential Monte Carlo methods: bsmc2(), cond_logLik(), eff_sample_size(), filter_mean(), filter_traj(), kalman, mif2(), pfilter(), pmcmc(), pred_var(), saved_states(), wpfilter()

Other extraction methods: coef(), cond_logLik(), covmat(), eff_sample_size(), filter_mean(), filter_traj(), forecast(), logLik, obs(), pred_var(), saved_states(), spy(), states(), summary(), timezero(), time(), traces()

[Package pomp version 5.2.5.1 Index]