ricker {pomp}	R Documentation

Ricker model with Poisson observations.

Description

ricker is a ‘pomp’ object encoding a stochastic Ricker model with Poisson measurement error.

Usage

ricker(r = exp(3.8), sigma = 0.3, phi = 10, c = 1, N_0 = 7)

Arguments

r
intrinsic growth rate

sigma
environmental process noise s.d.

phi
sampling rate

c
density dependence parameter

N_0
initial condition

Details

The state process is N_{t+1} = r N_{t} \exp(-c N_{t}+e_{t}), where the e_t are i.i.d. normal random deviates with zero mean and variance \sigma^2. The observed variables y_t are distributed as \mathrm{Poisson}(\phi N_t).

Value

A ‘pomp’ object containing the Ricker model and simulated data.

See Also

More examples provided with pomp: blowflies, childhood_disease_data, compartmental_models, dacca(), ebola, gompertz(), ou2(), pomp_examples, rw2(), verhulst()

Examples

po <- ricker()
plot(po)
coef(po)
simulate(po) |> plot()
 # takes too long for R CMD check
  ## generate a bifurcation diagram for the Ricker map
  p <- parmat(coef(ricker()),nrep=500)
  p["r",] <- exp(seq(from=1.5,to=4,length=500))
  trajectory(
    ricker(),
    times=seq(from=1000,to=2000,by=1),
    params=p,
    format="array"
  ) -> x
  matplot(p["r",],x["N",,],pch='.',col='black',
    xlab=expression(log(r)),ylab="N",log='x')

[Package pomp version 6.3.0.0 Index]