rmeasure {pomp}	R Documentation

rmeasure workhorse

Description

Sample from the measurement model distribution, given values of the latent states and the parameters.

Usage

## S4 method for signature 'pomp'
rmeasure(
  object,
  ...,
  x = states(object),
  times = time(object),
  params = coef(object)
)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

...
additional arguments are ignored.

x
an array containing states of the unobserved process. The dimensions of x are nvars x nrep x ntimes, where nvars is the number of state variables, nrep is the number of replicates, and ntimes is the length of times. One can also pass x as a named numeric vector, which is equivalent to the nrep=1, ntimes=1 case.

times
a numeric vector (length ntimes) containing times. These must be in non-decreasing order.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

Value

rmeasure returns a rank-3 array of dimensions nobs x nrep x ntimes, where nobs is the number of observed variables.

See Also

Specification of the measurement-model simulator: rmeasure_spec

More on pomp workhorse functions: dinit(), dmeasure(), dprior(), dprocess(), emeasure(), flow(), partrans(), pomp-package, rinit(), rprior(), rprocess(), skeleton(), vmeasure(), workhorses

[Package pomp version 6.1.0.1 Index]