rprior {pomp}	R Documentation

rprior workhorse

Description

Sample from the prior probability distribution.

Usage

## S4 method for signature 'pomp'
rprior(object, params = coef(object), ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
additional arguments are ignored.

Value

A numeric matrix containing the required samples.

See Also

Specification of the prior distribution simulator: prior_spec

More on pomp workhorse functions: dinit(), dmeasure(), dprior(), dprocess(), emeasure(), flow(), partrans(), pomp-package, rinit(), rmeasure(), rprocess(), skeleton(), vmeasure(), workhorses

More on Bayesian methods: abc(), bsmc2(), dprior(), pmcmc(), prior_spec

[Package pomp version 5.7.0.1 Index]