verhulst {pomp}	R Documentation

Verhulst-Pearl model

Description

The Verhulst-Pearl (logistic) model of population growth.

Usage

verhulst(n_0 = 10000, K = 10000, r = 0.9, sigma = 0.4, tau = 0.1, dt = 0.01)

Arguments

n_0
initial condition

K
carrying capacity

r
intrinsic growth rate

sigma
environmental process noise s.d.

tau
measurement error s.d.

dt
Euler timestep

Details

A stochastic version of the Verhulst-Pearl logistic model. This evolves in continuous time, according to the stochastic differential equation

dn_t = r\,n_t\,\left(1-\frac{n_t}{K}\right)\,dt+\sigma\,n_t\,dW_t.

Numerically, we simulate the stochastic dynamics using an Euler approximation.

The measurements are assumed to be log-normally distributed:

N_t \sim \mathrm{Lognormal}\left(\log{n_t},\tau\right).

Value

A ‘pomp’ object containing the model and simulated data. The following basic components are included in the ‘pomp’ object: ‘rinit’, ‘rprocess’, ‘rmeasure’, ‘dmeasure’, and ‘skeleton’.

See Also

More examples provided with pomp: blowflies, childhood_disease_data, compartmental_models, dacca(), ebola, gompertz(), ou2(), pomp_examples, ricker(), rw2()

Examples

 # takes too long for R CMD check
  verhulst() -> po
  plot(po)
  plot(simulate(po))
  pfilter(po,Np=1000) -> pf
  logLik(pf)
  spy(po)

[Package pomp version 5.7.0.1 Index]