vmeasure {pomp}	R Documentation

vmeasure

Description

Return the covariance matrix of the observed variables, given values of the latent states and the parameters.

Usage

## S4 method for signature 'pomp'
vmeasure(
  object,
  x = states(object),
  times = time(object),
  params = coef(object),
  ...
)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

x
an array containing states of the unobserved process. The dimensions of x are nvars x nrep x ntimes, where nvars is the number of state variables, nrep is the number of replicates, and ntimes is the length of times. One can also pass x as a named numeric vector, which is equivalent to the nrep=1, ntimes=1 case.

times
a numeric vector (length ntimes) containing times. These must be in non-decreasing order.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
additional arguments are ignored.

Value

vmeasure returns a rank-4 array of dimensions nobs x nobs x nrep x ntimes, where nobs is the number of observed variables. If v is the returned array, v[,,j,k] contains the covariance matrix at time times[k] given the state x[,j,k].

See Also

Specification of the measurement-model covariance matrix: vmeasure specification

More on pomp workhorse functions: dmeasure(), dprior(), dprocess(), emeasure(), flow(), partrans(), pomp-package, rinit(), rmeasure(), rprior(), rprocess(), skeleton(), workhorses

[Package pomp version 4.6.2.1 Index]