workhorses {pomp}	R Documentation

Workhorse functions for the pomp algorithms.

Description

These functions mediate the interface between the user's model and the package algorithms. They are low-level functions that do the work needed by the package's inference methods.

Details

They include

rinit

which samples from the initial-state distribution,

dinit

which evaluates the initial-state density,

dmeasure

which evaluates the measurement model density,

rmeasure

which samples from the measurement model distribution,

emeasure

which computes the expectation of the observed variables conditional on the latent state,

vmeasure

which computes the covariance matrix of the observed variables conditional on the latent state,

dprocess

which evaluates the process model density,

rprocess

which samples from the process model distribution,

dprior

which evaluates the prior probability density,

rprior

which samples from the prior distribution,

skeleton

which evaluates the model's deterministic skeleton,

flow

which iterates or integrates the deterministic skeleton to yield trajectories,

partrans

which performs parameter transformations associated with the model.

Author(s)

Aaron A. King

See Also

basic model components, elementary algorithms, estimation algorithms

More on pomp workhorse functions: dinit(), dmeasure(), dprior(), dprocess(), emeasure(), flow(), partrans(), pomp-package, rinit(), rmeasure(), rprior(), rprocess(), skeleton(), vmeasure()

[Package pomp version 5.7.0.1 Index]