Approximate Bayesian computation {pomp}	R Documentation

Estimation by approximate Bayesian computation (ABC)

Description

The approximate Bayesian computation (ABC) algorithm for estimating the parameters of a partially-observed Markov process.

Usage

## S4 method for signature 'pomp'
abc(object, Nabc = 1, start, proposal,
    pars, rw.sd, probes, scale, epsilon,
    verbose = getOption("verbose"), ...)
## S4 method for signature 'probed.pomp'
abc(object, probes,
    verbose = getOption("verbose"), ...)
## S4 method for signature 'abc'
abc(object, Nabc, start, proposal,
    probes, scale, epsilon,
    verbose = getOption("verbose"), ...)
## S4 method for signature 'abc'
continue(object, Nabc = 1, ...)
## S4 method for signature 'abc'
conv.rec(object, pars, ...)
## S4 method for signature 'abcList'
conv.rec(object, ...)
## S4 method for signature 'abc'
plot(x, y, pars, scatter = FALSE, ...)
## S4 method for signature 'abcList'
plot(x, y, ...)

Arguments

object
An object of class pomp.

Nabc
The number of ABC iterations to perform.

start
named numeric vector; the starting guess of the parameters.

proposal
optional function that draws from the proposal distribution. Currently, the proposal distribution must be symmetric for proper inference: it is the user's responsibility to ensure that it is. Several functions that construct appropriate proposal function are provided: see MCMC proposal functions for more information.

rw.sd
Deprecated. Will be removed in a future release. Specifying rw.sd is equivalent to setting proposal=mvn.diag.rw(rw.sd).

probes
List of probes (AKA summary statistics). See probe for details.

scale
named numeric vector of scales.

epsilon
ABC tolerance.

verbose
logical; if TRUE, print progress reports.

pars
Names of parameters.

scatter
optional logical; If TRUE, draw scatterplots. If FALSE, draw traceplots.

x
abc object.

y
Ignored.

...
Additional arguments. These are currently ignored.

Running ABC

abc returns an object of class abc. One or more abc objects can be joined to form an abcList object.

Re-running ABC iterations

To re-run a sequence of ABC iterations, one can use the abc method on a abc object. By default, the same parameters used for the original ABC run are re-used (except for tol, max.fail, and verbose, the defaults of which are shown above). If one does specify additional arguments, these will override the defaults.

Continuing ABC iterations

One can continue a series of ABC iterations from where one left off using the continue method. A call to abc to perform Nabc=m iterations followed by a call to continue to perform Nabc=n iterations will produce precisely the same effect as a single call to abc to perform Nabc=m+n iterations. By default, all the algorithmic parameters are the same as used in the original call to abc. Additional arguments will override the defaults.

Methods

Methods that can be used to manipulate, display, or extract information from an abc object:

conv.rec

conv.rec(object, pars) returns the columns of the convergence-record matrix corresponding to the names in pars. By default, all rows are returned.

c

Concatenates abc objects into an abcList.

plot

Diagnostic plots.

Author(s)

Edward L. Ionides ionides at umich dot edu, Aaron A. King kingaa at umich dot edu

References

T. Toni and M. P. H. Stumpf, Simulation-based model selection for dynamical systems in systems and population biology, Bioinformatics 26:104–110, 2010.

T. Toni, D. Welch, N. Strelkowa, A. Ipsen, and M. P. H. Stumpf, Approximate Bayesian computation scheme for parameter inference and model selection in dynamical systems Journal of the Royal Society, Interface 6:187–202, 2009.

See Also

pomp, probe, and the tutorials on the package website.

[Package pomp version 0.62-5 Index]