blowflies {pomp}	R Documentation

Nicholson's blowflies.

Description

blowflies is a data frame containing the data from several of Nicholson's classic experiments with the Australian sheep blowfly, Lucilia cuprina.

Usage

blowflies1(
  P = 3.2838,
  delta = 0.16073,
  N0 = 679.94,
  sigma.P = 1.3512,
  sigma.d = 0.74677,
  sigma.y = 0.026649
)

blowflies2(
  P = 2.7319,
  delta = 0.17377,
  N0 = 800.31,
  sigma.P = 1.442,
  sigma.d = 0.76033,
  sigma.y = 0.010846
)

Arguments

P
reproduction parameter

delta
death rate

N0
population scale factor

sigma.P
intensity of e noise

sigma.d
intensity of eps noise

sigma.y
measurement error s.d.

Details

blowflies1() and blowflies2() construct ‘pomp’ objects encoding stochastic delay-difference equation models. The data for these come from "population I", a control culture. The experiment is described on pp. 163–4 of Nicholson (1957). Unlimited quantities of larval food were provided; the adult food supply (ground liver) was constant at 0.4g per day. The data were taken from the table provided by Brillinger et al. (1980).

The models are discrete delay equations:

R[t+1] ~ Poisson(P N[t-tau] exp(-N[t-tau]/N0) e[t+1] dt)

S[t+1] ~ binomial(N[t],exp(-delta eps[t+1] dt))

N[t]=R[t]+S[t]

where e[t] and eps[t] are Gamma-distributed i.i.d. random variables with mean 1 and variances sigma.P^2/dt, sigma.d^2/dt, respectively. blowflies1 has a timestep (dt) of 1 day; blowflies2 has a timestep of 2 days. The process model in blowflies1 thus corresponds exactly to that studied by Wood (2010). The measurement model in both cases is taken to be

y[t] ~ negbin(N[t],1/sigma.y^2),

i.e., the observations are assumed to be negative-binomially distributed with mean N[t] and variance N[t]+(sigma.y N[t])^2.

Default parameter values are the MLEs as estimated by Ionides (2011).

Value

blowflies1 and blowflies2 return ‘pomp’ objects containing the actual data and two variants of the model.

References

A.J. Nicholson. The self-adjustment of populations to change. Cold Spring Harbor Symposia on Quantitative Biology 22, 153–173, 1957.

Y. Xia and H. Tong. Feature matching in time series modeling. Statistical Science 26, 21–46, 2011.

E.L. Ionides. Discussion of “Feature matching in time series modeling” by Y. Xia and H. Tong. Statistical Science 26, 49–52, 2011.

S. N. Wood Statistical inference for noisy nonlinear ecological dynamic systems. Nature 466, 1102–1104, 2010.

W.S.C. Gurney, S.P. Blythe, and R.M. Nisbet. Nicholson's blowflies revisited. Nature 287, 17–21, 1980.

D.R. Brillinger, J. Guckenheimer, P. Guttorp, and G. Oster. Empirical modelling of population time series: The case of age and density dependent rates. In: G. Oster (ed.), Some Questions in Mathematical Biology vol. 13, pp. 65–90, American Mathematical Society, Providence, 1980.

See Also

More examples provided with pomp: bsflu, dacca(), ebola, gompertz(), measles, ou2(), parus, pomp_examples, ricker(), rw2(), sir_models, verhulst()

More data sets provided with pomp: bsflu, dacca(), ebola, measles, parus

Examples

plot(blowflies1())
plot(blowflies2())

[Package pomp version 3.5.1.0 Index]