coef {pomp}	R Documentation

Extract, set, or alter coefficients

Description

Extract, set, or modify the estimated parameters from a fitted model.

Usage

## S4 method for signature 'listie'
coef(object, ...)

## S4 method for signature 'pomp'
coef(object, pars, transform = FALSE, ...)

## S4 replacement method for signature 'pomp'
coef(object, pars, transform = FALSE, ...) <- value

## S4 method for signature 'objfun'
coef(object, ...)

Arguments

object
an object of class ‘pomp’, or of a class extending ‘pomp’

...
ignored

pars
optional character; names of parameters to be retrieved or set.

transform
logical; perform parameter transformation?

value
numeric vector or list; values to be assigned. If value = NULL, the parameters are unset.

Details

coef allows one to extract the parameters from a fitted model.

coef(object,transform=TRUE) returns the parameters transformed onto the estimation scale.

coef(object) <- value sets or alters the coefficients of a ‘pomp’ object.

coef(object,transform=TRUE) <- value assumes that value is on the estimation scale, and applies the “from estimation scale” parameter transformation from object before altering the coefficients.

[Package pomp version 3.5.0.0 Index]