continue {pomp}	R Documentation

Continue an iterative calculation

Description

Continue an iterative computation where it left off.

Usage

continue(object, ...)

## S4 method for signature 'abcd_pomp'
continue(object, Nabc = 1, ...)

## S4 method for signature 'pmcmcd_pomp'
continue(object, Nmcmc = 1, ...)

## S4 method for signature 'mif2d_pomp'
continue(object, Nmif = 1, ...)

Arguments

object
the result of an iterative pomp computation

...
additional arguments will be passed to the underlying method. This allows one to modify parameters used in the original computations.

Nabc
positive integer; number of additional ABC iterations to perform

Nmcmc
positive integer; number of additional PMCMC iterations to perform

Nmif
positive integer; number of additional filtering iterations to perform

See Also

mif2 pmcmc abc

[Package pomp version 3.3.0.1 Index]