Csnippet {pomp}	R Documentation

C code snippets for accelerating computations

Description

For including snippets of C code in pomp objects.

Usage

  Csnippet(text)

Arguments

text
character; a snippet of C code.

Value

An object of class Csnippet.

Using C snippets to accelerate computations

From version 0.50, pomp provides a facility whereby users can define their model's components using inline C code. Furnishing one or more Csnippets as arguments to the pomp constructor causes them to be written to a C file stored in the R session's temporary directory, which is then compiled (via R CMD SHLIB) into a dynamically loadable shared object file. This is then loaded as needed.

Note to Windows and Mac users: By default, your R installation may not support R CMD SHLIB. Have a look at the tutorials on the package website for instructions on enabling this powerful feature of R.

In writing a Csnippet one must bear in mind both the goal of the snippet, i.e., what computation it is intended to perform, and the context in which it will be executed. Details of both of these are given below in the form of rules governing the use of Csnippets. Illustrative examples are given in the tutorials on the package website.

General rules for writing C snippets

C snippets must be valid C. They will embedded verbatim in a template file which will then be compiled by a call to R CMD SHLIB. If the resulting file does not compile, an error message wil be generated. No attempt is made by pomp to interpret this message. Typically, compilation errors are due to either invalid C syntax or undeclared variables.

State variables, parameters, observables, and covariates must be left undeclared in the snippet. State variables and parameters must be declared in either the statenames or paramnames argument to pomp, as appropriate. Compiler errors that complain about undeclared state variables or parameters are usually due to failure to include these parameters in the appropriate vector of names.

A C snippet can declare local variables. Be careful not to use names that match those of state variables, observables, or parameters. The latter must never be declared within a snippet.

Names of observables are determined by their names in the data. They must be referred to in measurement model snippets (rmeasure or dmeasure) by those names.

If the pomp object contains a table of covariates (see pomp), then the variables in the covariate table will be available, by their names, in the context within which the snippet is executed.

R variables with names containing dots (‘.’) are replaced in the C codes by variable names in which all dots have been replaced by underscores (‘_’).

The header ‘R.h’, provided with R, will be included in the generated C file, making all of the R C API available for use in the snippet.

The header ‘pomp.h’, provided with pomp, will also be included, making all of the pomp C API available for use in every snippet. Do

file.show(system.file("include/pomp.h",package="pomp"))
to view this header file.

Snippets of C code passed to the globals argument of pomp will be included at the head of the generated C file. This can be used to declare global variables, define useful functions, and include arbitrary header files.

Rules for writing rmeasure snippets

The goal of such a snippet is to fill the observables with random values drawn from the measurement model distribution. Accordingly, each observable should be assigned a new value.

In addition to the states, parameters, covariates (if any), and observables, the variable t, containing the time of the observation, will be defined in the context in which the snippet is executed.

Rules for writing dmeasure snippets

The goal of such a snippet is to fill the lik variable with the likelihood of the data given the state. Alternatively, if give_log=1, lik should be filled with the log likelihood.

In addition to the states, parameters, covariates (if any), and observables, the variable t, containing the time of the observation, will be defined in the context in which the snippet is executed.

Rules for writing euler.sim and discrete.time.sim snippets

The goal of such a snippet is to replace the state variables with their new random values at the end of the time interval. Accordingly, each state variable should be over-written with its new value.

In addition to the states, parameters, covariates (if any), and observables, the variables t and dt, containing respectively the time at the beginning of the Euler step and the Euler step's duration, will be defined in the context in which the snippet is executed.

Rules for writing skeleton snippets

For each state variable, there is a corresponding component of the deterministic skeleton. The goal of such a snippet is to compute all the components.

When the skeleton is a map, the component corresponding to state variable x is named Dx and is the new value of x after one iteration of the map.

When the skeleton is a vectorfield, the component corresponding to state variable x is named Dx and is the value of $dx/dt$.

As with the other C snippets, all states, parameters and covariates, as well as the current time, t, will be defined in the context within which the snippet is executed.

Rules for writing parameter transformation snippets

The parameter transformation mapping a parameter vector from the scale defined by the model codes to another scale is specified as the parameter.inv.transform whilst the transformation mapping a parameter vector from the alternative scale to that on which the model is defined is specified as the parameter.transform.

The goal of a these snippets is the computation of the values of the transformed parameters. The value of transformed parameter x should be assigned to variable Tx.

Time-, state-, and covariate-dependent transformations are not allowed. Therefore, neither the time, nor any state variables, nor any of the covariates will be available in the context within which a parameter transformation snippet is executed.

Viewing the generated C code

It can be useful to view the C code generated by calling pomp with one or more Csnippet arguments. To do so, set options(verbose=TRUE) before calling pomp. A message giving the name of the generated C file (in the session temporary directory) will be printed.

Author(s)

Aaron A. King kingaa at umich dot edu

See Also

pomp, plugins, and the tutorials on the package website.

[Package pomp version 0.62-5 Index]