dprior {pomp}	R Documentation

dprior

Description

Evaluates the prior probability density.

Usage

## S4 method for signature 'pomp'
dprior(object, params, ..., log = FALSE)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
additional arguments are ignored.

log
if TRUE, log probabilities are returned.

Value

The required density (or log density), as a numeric vector.

See Also

Specification of the prior density evaluator: prior_spec

More on pomp workhorse functions: dmeasure(), dprocess(), flow(), partrans(), rinit(), rmeasure(), rprior(), rprocess(), skeleton(), workhorses

[Package pomp version 3.4.5.0 Index]