ebola {pomp}	R Documentation

Ebola outbreak, West Africa, 2014-2016

Description

Data and models for the 2014–2016 outbreak of Ebola virus disease in West Africa.

Usage

ebolaModel(
  country = c("GIN", "LBR", "SLE"),
  data = NULL,
  timestep = 1/8,
  nstageE = 3L,
  R0 = 1.4,
  rho = 0.2,
  cfr = 0.7,
  k = 0,
  index_case = 10,
  incubation_period = 11.4,
  infectious_period = 7
)

Arguments

country
ISO symbol for the country (GIN=Guinea, LBR=Liberia, SLE=Sierra Leone).

data
if NULL, the situation report data (see ebolaWHO) for the appropriate country or region will be used. Providing a dataset here will override this behavior.

timestep
duration (in days) of Euler time-step for the simulations.

nstageE
integer; number of incubation stages.

R0
basic reproduction ratio

rho
case reporting efficiency

cfr
case fatality rate

k
dispersion parameter (negative binomial size parameter)

index_case
number of cases on day 0 (2014-04-01)

incubation_period, infectious_period
mean duration (in days) of the incubation and infectious periods.

Details

The data include monthly case counts and death reports derived from WHO situation reports, as reported by the U.S. CDC. The models are described in King et al. (2015).

The data-cleaning script is included in the R source code file ‘ebola.R’.

Model structure

The default incubation period is supposed to be Gamma distributed with shape parameter nstageE and mean 11.4 days and the case-fatality ratio ('cfr') is taken to be 0.7 (cf. WHO Ebola Response Team 2014). The discrete-time formula is used to calculate the corresponding alpha (cf. He et al. 2010).

The observation model is a hierarchical model for cases and deaths:

p(R_t, D_t| C_t) = p(R_t | C_t) p(D_t | C_t, R_t).

Here, p(R_t | C_t) is negative binomial with mean ρ C_t and dispersion parameter 1/k; p(D_t | C_t, R_t) is binomial with size R_t and probability equal to the case fatality rate cfr.

References

A.A. King, M. Domenech de Cellès, F.M.G. Magpantay, and P. Rohani. Avoidable errors in the modelling of outbreaks of emerging pathogens, with special reference to Ebola. Proceedings of the Royal Society of London, Series B 282, 20150347, 2015.

WHO Ebola Response Team. Ebola virus disease in West Africa—the first 9 months of the epidemic and forward projections. New England Journal of Medicine 371, 1481–1495, 2014.

D. He, E.L. Ionides, & A.A. King. Plug-and-play inference for disease dynamics: measles in large and small populations as a case study. Journal of the Royal Society Interface 7, 271–283, 2010.

See Also

More data sets provided with pomp: blowflies, bsflu, dacca(), measles, parus

More examples provided with pomp: blowflies, bsflu, dacca(), gompertz(), measles, ou2(), parus, pomp_examples, ricker(), rw2(), sir_models, verhulst()

Examples

data(ebolaWA2014)

library(ggplot2)
library(tidyr)

ebolaWA2014 %>%
  gather(variable,count,cases,deaths) %>%
  ggplot(aes(x=date,y=count,group=country,color=country))+
  geom_line()+
  facet_grid(variable~.,scales="free_y")+
  theme_bw()+
  theme(axis.text=element_text(angle=-90))

ebolaWA2014 %>%
  gather(variable,count,cases,deaths) %>%
  ggplot(aes(x=date,y=count,group=variable,color=variable))+
  geom_line()+
  facet_grid(country~.,scales="free_y")+
  theme_bw()+
  theme(axis.text=element_text(angle=-90))

plot(ebolaModel(country="SLE"))
plot(ebolaModel(country="LBR"))
plot(ebolaModel(country="GIN"))

[Package pomp version 3.1.1.9 Index]