eff.sample.size {pomp}	R Documentation

Effective sample size

Description

Estimate the effective sample size of a Monte Carlo computation.

Usage

## S4 method for signature 'bsmcd_pomp'
eff.sample.size(object, ...)

## S4 method for signature 'pfilterd_pomp'
eff.sample.size(object, ...)

## S4 method for signature 'wpfilterd_pomp'
eff.sample.size(object, ...)

Arguments

object
result of a filtering computation

...
ignored

Details

Effective sample size is computed as

1/(sum(w_it^2)),

where w_it is the normalized weight of particle i at time t.

See Also

More on particle-filter based methods in pomp: bsmc2(), cond.logLik(), filter.mean(), filter.traj(), kalman, mif2(), pfilter(), pmcmc(), pred.mean(), pred.var(), saved.states(), wpfilter()

[Package pomp version 3.5.0.3 Index]