eff.sample.size {pomp}	R Documentation

Effective sample size

Description

Estimate the effective sample size of a Monte Carlo computation.

Usage

## S4 method for signature 'bsmcd_pomp'
eff.sample.size(object, ...)

## S4 method for signature 'pfilterd_pomp'
eff.sample.size(object, ...)

Arguments

object
result of a filtering computation

...
ignored

Details

Effective sample size is computed as

1/(sum(w_it^2)),

where w_it is the normalized weight of particle i at time t.

See Also

Other particle filter methods: bsmc2(), cond.logLik(), filter.mean(), filter.traj(), mif2(), pfilter(), pmcmc(), pred.mean(), pred.var()

[Package pomp version 2.7.1.0 Index]