elementary_algorithms {pomp}	R Documentation

Elementary algorithms.

Description

In pomp, elementary algorithms perform POMP model operations. These operations do not themselves estimate parameters, though they may be instrumental in inference methods.

Details

There are six elementary algorithms in pomp:

simulate which simulates from the joint distribution of latent and observed variables,

pfilter, which performs a simple particle filter operation,

wpfilter, which performs a weighted particle filter operation,

probe, which computes a suite of user-specified summary statistics to actual and simulated data,

spect, which performs a power-spectral density function computation on actual and simulated data,

trajectory, which iterates or integrates the deterministic skeleton (according to whether the latter is a (discrete-time) map or a (continuous-time) vectorfield.

Help pages detailing each elementary algorithm component are provided.

See Also

basic model components, workhorse functions, estimation algorithms.

More on pomp elementary algorithms: pfilter(), pomp-package, probe(), simulate(), spect(), trajectory(), wpfilter()

[Package pomp version 3.3.3.1 Index]