filter.mean {pomp}	R Documentation

Filtering mean

Description

The mean of the filtering distribution

Usage

## S4 method for signature 'kalmand_pomp'
filter.mean(object, vars, ...)

## S4 method for signature 'pfilterd_pomp'
filter.mean(object, vars, ...)

Arguments

object
result of a filtering computation

vars
optional character; names of variables

...
ignored

Details

The filtering distribution is that of

Xk | Y1=y1*,…,Yk=yk*,

where Xk, Yk are the latent state and observable processes, respectively, and yt* is the data, at time tk.

The filtering mean is therefore the expectation of this distribution

E[Xk | Y1=y1*,…,Yk=yk*].

See Also

More on particle-filter based methods in pomp: bsmc2(), cond.logLik(), eff.sample.size(), filter.traj(), kalman, mif2(), pfilter(), pmcmc(), pred.mean(), pred.var(), saved.states(), wpfilter()

[Package pomp version 3.4.6.0 Index]