flow {pomp}	R Documentation

Flow of a deterministic model

Description

Compute the flow induced by a deterministic vectorfield or map.

Usage

## S4 method for signature 'pomp'
flow(object, x0, t0, times, params, ..., verbose = getOption("verbose", FALSE))

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

x0
an array with dimensions nvar x nrep giving the initial conditions of the trajectories to be computed.

t0
the time at which the initial conditions are assumed to hold.

times
a numeric vector (length ntimes) containing times at which the itineraries are desired. These must be in non-decreasing order with times[1]>t0.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
Additional arguments are passed to the ODE integrator (if the skeleton is a vectorfield) and are ignored if it is a map. See ode for a description of the additional arguments accepted by the ODE integrator.

verbose
logical; if TRUE, diagnostic messages will be printed to the console.

Details

In the case of a discrete-time system (map), flow iterates the map to yield trajectories of the system. In the case of a continuous-time system (vectorfield), flow uses the numerical solvers in deSolve to integrate the vectorfield starting from given initial conditions.

Value

flow returns an array of dimensions nvar x nrep x ntimes. If x is the returned matrix, x[i,j,k] is the i-th component of the state vector at time times[k] given parameters params[,j].

See Also

skeleton, trajectory, rprocess

More on pomp workhorse functions: dmeasure(), dprior(), dprocess(), partrans(), rinit(), rmeasure(), rprior(), rprocess(), skeleton(), workhorses

[Package pomp version 3.4.1.0 Index]