gompertz {pomp}	R Documentation

Gompertz model with log-normal observations.

Description

gompertz() constructs a ‘pomp’ object encoding a stochastic Gompertz population model with log-normal measurement error.

Usage

gompertz(
  K = 1,
  r = 0.1,
  sigma = 0.1,
  tau = 0.1,
  X_0 = 1,
  times = 1:100,
  t0 = 0
)

Arguments

K
carrying capacity

r
growth rate

sigma
process noise intensity

tau
measurement error s.d.

X_0
value of the latent state variable X at the zero time

times
observation times

t0
zero time

Details

The state process is X[t+1]=K^(1-S) X[t]^S eps[t], where S=e^{-r} and the eps[t] are i.i.d. lognormal random deviates with variance sigma^2. The observed variables Y_t are distributed as lognormal(log(X[t]),tau). Parameters include the per-capita growth rate r, the carrying capacity K, the process noise s.d. sigma, the measurement error s.d. tau, and the initial condition X[0]. The ‘pomp’ object includes parameter transformations that log-transform the parameters for estimation purposes.

Value

A ‘pomp’ object with simulated data.

See Also

More examples provided with pomp: blowflies, bsflu, dacca(), ebola, measles, ou2(), parus, pomp_examples, ricker(), rw2(), sir_models, verhulst()

Examples

plot(gompertz())
plot(gompertz(K=2,r=0.01))

[Package pomp version 3.5.1.0 Index]