load {pomp}	R Documentation

Loading and unloading shared-object libraries

Description

pompLoad and pompUnload cause compiled codes associated with object to be dynamically linked or unlinked, respectively. solibs<- is a helper function for developers of packages that extend pomp.

Usage

## S4 method for signature 'pomp'
pompLoad(object, ...)

## S4 method for signature 'pomp'
pompUnload(object, ...)

## S4 replacement method for signature 'pomp'
solibs(object, ...) <- value

Arguments

object
an object of class ‘pomp’, or extending this class.

Details

When C snippets are used in the construction of a ‘pomp’ object, the resulting shared-object library is dynamically loaded (linked) before each use, and unloaded afterward.

solibs<- prepends the ‘lib’ generated by hitch to the ‘solibs’ slot of a ‘pomp’ object.

Value

NULL, invisibly.

[Package pomp version 2.8.0.3 Index]