logmeanexp {pomp}	R Documentation

The log-mean-exp trick

Description

logmeanexp computes

log mean exp(x_i),

avoiding over- and under-flow in doing so. It can optionally return an estimate of the standard error in this quantity.

Usage

logmeanexp(x, se = FALSE)

Arguments

x
numeric

se
logical; give approximate standard error?

Details

When se = TRUE, logmeanexp uses a jackknife estimate of the variance in log(x).

Value

log(mean(exp(x))) computed so as to avoid over- or underflow. If se = FALSE, the approximate standard error is returned as well.

Author(s)

Aaron A. King

Examples

## an estimate of the log likelihood:
po <- ricker()
ll <- replicate(n=5,logLik(pfilter(po,Np=1000)))
logmeanexp(ll)
## with standard error:
logmeanexp(ll,se=TRUE)

[Package pomp version 3.3.1.0 Index]