mcap {pomp}	R Documentation

Monte Carlo adjusted profile

Description

Given a collection of points maximizing the likelihood over a range of fixed values of a focal parameter, this function constructs a profile likelihood confidence interval accommodating both Monte Carlo error in the profile and statistical uncertainty present in the likelihood function.

Usage

mcap(logLik, parameter, level = 0.95, span = 0.75, Ngrid = 1000)

Arguments

logLik
numeric; a vector of profile log likelihood evaluations.

parameter
numeric; the corresponding values of the focal parameter.

level
numeric; the confidence level required.

span
numeric; the loess smoothing parameter.

Ngrid
integer; the number of points to evaluate the smoothed profile.

Value

mcap returns a list including the loess-smoothed profile, a quadratic approximation, and the constructed confidence interval.

Author(s)

Edward L. Ionides

References

E. L. Ionides, C. Bretó, J. Park, R. A. Smith, and A. A. King. Monte Carlo profile confidence intervals for dynamic systems. Journal of the Royal Society, Interface 14, 20170126, 2017. doi:10.1098/rsif.2017.0126.

[Package pomp version 5.11.1.0 Index]