partrans {pomp}	R Documentation

partrans

Description

Performs parameter transformations.

Usage

## S4 method for signature 'pomp'
partrans(object, params, dir = c("fromEst", "toEst"), ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

dir
the direction of the transformation to perform.

...
additional arguments are ignored.

Value

If dir=fromEst, the parameters in params are assumed to be on the estimation scale and are transformed onto the natural scale. If dir=toEst, they are transformed onto the estimation scale. In both cases, the parameters are returned as a named numeric vector or an array with rownames, as appropriate.

See Also

Specification of parameter transformations: parameter_trans

Other pomp workhorses: dmeasure(), dprior(), dprocess(), flow(), rinit(), rmeasure(), rprior(), rprocess(), skeleton(), workhorses

[Package pomp version 3.1.1.1 Index]