plot {pomp}	R Documentation

Plotting

Description

Diagnostic plots.

Usage

## S4 method for signature 'pomp_plottable'
plot(
  x,
  variables,
  panel = lines,
  nc = NULL,
  yax.flip = FALSE,
  mar = c(0, 5.1, 0, if (yax.flip) 5.1 else 2.1),
  oma = c(6, 0, 5, 0),
  axes = TRUE,
  ...
)

## S4 method for signature 'Pmcmc'
plot(x, ..., pars)

## S4 method for signature 'Abc'
plot(x, ..., pars, scatter = FALSE)

## S4 method for signature 'Mif2'
plot(x, y, ...)

## S4 method for signature 'probed_pomp'
plot(x, y, ...)

## S4 method for signature 'spectd_pomp'
plot(
  x,
  ...,
  max.plots.per.page = 4,
  plot.data = TRUE,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  quantile.styles = list(lwd = 1, lty = 1, col = "gray70"),
  data.styles = list(lwd = 2, lty = 2, col = "black")
)

## S4 method for signature 'bsmcd_pomp'
plot(x, pars, thin, ...)

## S4 method for signature 'probe_match_objfun'
plot(x, y, ...)

## S4 method for signature 'spect_match_objfun'
plot(x, y, ...)

Arguments

x
the object to plot

variables
optional character; names of variables to be displayed

panel
function of prototype panel(x, col, bg, pch, type, ...) which gives the action to be carried out in each panel of the display.

nc
the number of columns to use. Defaults to 1 for up to 4 series, otherwise to 2.

yax.flip
logical; if TRUE, the y-axis (ticks and numbering) should flip from side 2 (left) to 4 (right) from series to series.

mar, oma
the par mar and oma settings. Modify with care!

axes
logical; indicates if x- and y- axes should be drawn

...
ignored or passed to low-level plotting functions

pars
names of parameters.

scatter
logical; if FALSE, traces of the parameters named in pars will be plotted against ABC iteration number. If TRUE, the traces will be displayed or as a scatterplot.

y
ignored

max.plots.per.page
positive integer; maximum number of plots on a page

plot.data
logical; should the data spectrum be included?

quantiles
numeric; quantiles to display

quantile.styles
list; plot styles to use for quantiles

data.styles
list; plot styles to use for data

thin
integer; when the number of samples is very large, it can be helpful to plot a random subsample: thin specifies the size of this subsample.

[Package pomp version 3.1.1.9 Index]