pomp-package {pomp}	R Documentation

Inference for partially observed Markov processes

Description

The pomp package provides facilities for inference on time series data using partially-observed Markov process (POMP) models. These models are also known as state-space models or nonlinear stochastic dynamical systems. One can use pomp to fit nonlinear, non-Gaussian dynamic models to time-series data. The package is both a set of tools for data analysis and a platform upon which statistical inference methods for POMP models can be implemented.

Data analysis using pomp

The first step in using pomp is to encode one's model(s) and data in objects of class pomp. One does this via a call to pomp, which involves specifying the unobserved state process and the measurement process of the model. Details on this are given in the documentation for the pomp constructor function. Examples are given in the tutorials on the package website, in the demos (demo(package='pomp')), and via the pompExample function.

pomp version 0.62-5 provides algorithms for

simulation of stochastic dynamical systems; see simulate

particle filtering (AKA sequential Monte Carlo or sequential importance sampling); see pfilter

the iterated filtering method of Ionides et al. (2006, 2011, 2015); see mif

the nonlinear forecasting algorithm of Kendall et al. (2005); see nlf

the particle MCMC approach of Andrieu et al. (2010); see pmcmc

the probe-matching method of Kendall et al. (1999, 2005); see probe.match

a spectral probe-matching method (Reuman et al. 2006, 2008); see spect.match

synthetic likelihood a la Wood (2010); see probe

approximate Bayesian computation (Toni et al. 2009); see abc

the approximate Bayesian sequential Monte Carlo scheme of Liu & West (2001); see bsmc

simple trajectory matching; see traj.match.

The package also provides various tools for plotting and extracting information on models and data.

Developing inference tools on the pomp platform

pomp provides a very general interface to the components of POMP models. All the inference algorithms in pomp interact with the models and data via this interface. One goal of the pomp project has been to facilitate the development of new algorithms in an environment where they can be tested and compared on a growing body of models and datasets.

The low-level interface relevant to developers is documented here.

Comments, bug reports, feature requests

Contributions are welcome, as are comments, feature requests, and bug reports. See the package website http://pomp.r-forge.r-project.org for more information, access to the package mailing list, links to the authors' websites, references to the literature, and up-to-date versions of the package source and documentation.

Documentation

A number of tutorials, demonstrating the construction of pomp objects and the application of various inference algorithms, are available on the package homepage: http://pomp.r-forge.r-project.org. Several examples of the construction of pomp objects are provided with the package and are documented in the help pages: to view a full list of these, execute pompExample(). In addition, there are a number of demos, which can be viewed by executing demo(package="pomp").

History

Much of the groundwork for pomp was laid by a working group of the National Center for Ecological Analysis and Synthesis (NCEAS), “Inference for Mechanistic Models”.

License

pomp is provided under the GNU Public License (GPL).

Author(s)

Aaron A. King kingaa at umich dot edu

References

See the package website, http://pomp.r-forge.r-project.org, for the references.

See Also

pomp, pomp low-level interface, pfilter, simulate, mif, nlf, probe, traj.match, bsmc2, pmcmc

[Package pomp version 0.62-5 Index]