pomp-package {pomp}	R Documentation

Inference for partially observed Markov processes

Description

The pomp package provides facilities for inference on time series data using partially-observed Markov process (POMP) models. These models are also known as state-space models, hidden Markov models, or nonlinear stochastic dynamical systems. One can use pomp to fit nonlinear, non-Gaussian dynamic models to time-series data. The package is both a set of tools for data analysis and a platform upon which statistical inference methods for POMP models can be implemented.

Data analysis using pomp

pomp provides algorithms for

simulation of stochastic dynamical systems; see simulate

particle filtering (AKA sequential Monte Carlo or sequential importance sampling); see pfilter

the iterated filtering methods of Ionides et al. (2006, 2011, 2015); see mif2

the nonlinear forecasting algorithm of Kendall et al. (2005); see nlf

the particle MCMC approach of Andrieu et al. (2010); see pmcmc

the probe-matching method of Kendall et al. (1999, 2005); see probe.match

a spectral probe-matching method (Reuman et al. 2006, 2008); see spect.match

synthetic likelihood a la Wood (2010); see probe

approximate Bayesian computation (Toni et al. 2009); see abc

the approximate Bayesian sequential Monte Carlo scheme of Liu & West (2001); see bsmc2

ensemble and ensemble adjusted Kalman filters; see kalman

simple trajectory matching; see traj.match.

The package also provides various tools for plotting and extracting information on models and data.

Author(s)

Aaron A. King

References

A. A. King, D. Nguyen, and E. L. Ionides. Statistical Inference for Partially Observed Markov Processes via the Package pomp. Journal of Statistical Software 69(12), 1–43, 2016. An updated version of this paper is available on the package website.

See the package website, https://kingaa.github.io/pomp/, for more references.

See Also

Other information on model implementation: Csnippet, accumulators, covariate_table(), distributions, dmeasure_spec, dprocess_spec, parameter_trans(), prior_spec, rinit_spec, rmeasure_spec, rprocess_spec, skeleton_spec, transformations, userdata

Other pomp parameter estimation methods: abc(), bsmc2(), kalman, mif2(), nlf, pmcmc(), probe.match, spect.match

Other elementary POMP methods: pfilter(), probe(), simulate(), spect(), wpfilter()

[Package pomp version 3.0.1.0 Index]