pomp_fun {pomp}	R Documentation

The "pomp_fun" class

Description

Definition and methods of the ‘pomp_fun’ class.

Usage

## S4 method for signature 'missing'
pomp_fun(
  slotname = NULL,
  obsnames = character(0),
  statenames = character(0),
  paramnames = character(0),
  covarnames = character(0),
  ...
)

## S4 method for signature ''function''
pomp_fun(f, proto = NULL, slotname = NULL, ...)

## S4 method for signature 'character'
pomp_fun(
  f,
  PACKAGE = NULL,
  obsnames = character(0),
  statenames = character(0),
  paramnames = character(0),
  covarnames = character(0),
  slotname = NULL,
  ...
)

## S4 method for signature 'Csnippet'
pomp_fun(
  f,
  slotname = NULL,
  libname = NULL,
  obsnames = character(0),
  statenames = character(0),
  paramnames = character(0),
  covarnames = character(0),
  Cname,
  ...
)

## S4 method for signature 'pomp_fun'
pomp_fun(f, ...)

Arguments

f
A function or the name of a native routine.

proto
optional string; a prototype against which f will be checked.

PACKAGE
optional; the name of the dynamically-loadable library in which the native function f can be found.

object, x
the ‘pomp_fun’ object.

Details

The ‘pomp_fun’ class implements a common interface for user-defined procedures that can be defined in terms of R code or by compiled native routines.

Author(s)

Aaron A. King

See Also

pomp

[Package pomp version 3.0.2.1 Index]