pred.mean {pomp}	R Documentation

Prediction mean

Description

The mean of the prediction distribution

Usage

## S4 method for signature 'kalmand_pomp'
pred.mean(object, vars, ...)

## S4 method for signature 'pfilterd_pomp'
pred.mean(object, vars, ...)

Arguments

object
result of a filtering computation

vars
optional character; names of variables

...
ignored

Details

The prediction distribution is that of

Xk | Y1=y1*,…,Y(k-1)=y(k-1)*,

where Xk, Yk are the latent state and observable processes, respectively, and yk* is the data, at time tk.

The prediction mean is therefore the expectation of this distribution

E[Xk | Y1=y1*,…,Y(k-1)=y(k-1)*].

See Also

More on particle-filter based methods in pomp: bsmc2(), cond.logLik(), eff.sample.size(), filter.mean(), filter.traj(), kalman, mif2(), pfilter(), pmcmc(), pred.var(), saved.states(), wpfilter()

[Package pomp version 3.1.1.12 Index]