proposals {pomp}	R Documentation

MCMC proposal distributions

Description

Functions to construct proposal distributions for use with MCMC methods.

Usage

mvn.diag.rw(rw.sd)

mvn.rw(rw.var)

mvn.rw.adaptive(
  rw.sd,
  rw.var,
  scale.start = NA,
  scale.cooling = 0.999,
  shape.start = NA,
  target = 0.234,
  max.scaling = 50
)

Arguments

rw.sd
named numeric vector; random-walk SDs for a multivariate normal random-walk proposal with diagonal variance-covariance matrix.

rw.var
square numeric matrix with row- and column-names. Specifies the variance-covariance matrix for a multivariate normal random-walk proposal distribution.

scale.start, scale.cooling, shape.start, target, max.scaling
parameters to control the proposal adaptation algorithm. Beginning with MCMC iteration scale.start, the scale of the proposal covariance matrix will be adjusted in an effort to match the target acceptance ratio. This initial scale adjustment is “cooled”, i.e., the adjustment diminishes as the chain moves along. The parameter scale.cooling specifies the cooling schedule: at n iterations after scale.start, the current scaling factor is multiplied with scale.cooling^n. The maximum scaling factor allowed at any one iteration is max.scaling. After shape.start accepted proposals have accumulated, a scaled empirical covariance matrix will be used for the proposals, following Roberts and Rosenthal (2009).

Value

Each of these calls constructs a function suitable for use as the proposal argument of pmcmc or abc. Given a parameter vector, each such function returns a single draw from the corresponding proposal distribution.

Author(s)

Aaron A. King, Sebastian Funk

References

G.O. Roberts and J.S. Rosenthal. Examples of Adaptive MCMC. Journal of Computational and Graphical Statistics 18, 349–367, 2009.

See Also

pmcmc, abc

[Package pomp version 3.1.1.10 Index]