MCMC proposal distributions {pomp}	R Documentation

MCMC proposal distributions

Description

Functions to construct proposal distributions for use with MCMC methods.

Usage

  mvn.diag.rw(rw.sd)
  mvn.rw(rw.var)

Arguments

rw.sd
named numeric vector; random-walk SDs for a multivariate normal random-walk proposal with diagonal variance-covariance matrix.

rw.var
square numeric matrix with row- and column-names. Specifies the variance-covariance matrix for a multivariate normal random-walk proposal distribution.

Value

Functions taking a single argument. Given a parameter vector, each returns a single draw from the corresponding proposal distribution.

Author(s)

Aaron A. King kingaa at umich dot edu

See Also

pmcmc, abc

[Package pomp version 0.62-5 Index]