ricker {pomp}	R Documentation

Ricker model with Poisson observations.

Description

ricker is a ‘pomp’ object encoding a stochastic Ricker model with Poisson measurement error.

Usage

ricker(r = exp(3.8), sigma = 0.3, phi = 10, c = 1, N_0 = 7)

Arguments

r
intrinsic growth rate

sigma
environmental process noise s.d.

phi
sampling rate

c
density dependence parameter

N_0
initial condition

Details

The state process is N[t+1] = r N[t] exp(-c N[t]+e[t]), where the e[t] are i.i.d. normal random deviates with zero mean and variance sigma^2. The observed variables y[t] are distributed as Poisson(phi N[t]).

Value

A ‘pomp’ object containing the Ricker model and simulated data.

See Also

Other pomp examples: blowflies, bsflu, dacca(), ebola, gompertz(), measles, ou2(), parus, pomp_examples, rw2(), sir_models, verhulst()

Examples

po <- ricker()
plot(po)
coef(po)
simulate(po) %>% plot()

[Package pomp version 3.0.0.0 Index]