rinit {pomp}	R Documentation

rinit

Description

Samples from the initial-state distribution.

Usage

## S4 method for signature 'pomp'
rinit(object, params, t0, nsim = 1, ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

t0
the initial time, i.e., the time corresponding to the initial-state distribution.

nsim
optional integer; the number of initial states to simulate per column of params.

...
additional arguments are ignored.

Value

rinit returns an nvar x nsim*ncol(params) matrix of state-process initial conditions when given an npar x nsim matrix of parameters, params, and an initial time t0. By default, t0 is the initial time defined when the ‘pomp’ object ws constructed.

See Also

Specification of the initial-state distribution: rinit_spec

More on pomp workhorse functions: dmeasure(), dprior(), dprocess(), flow(), partrans(), rmeasure(), rprior(), rprocess(), skeleton(), workhorses

[Package pomp version 3.3.0.0 Index]