rprior {pomp}	R Documentation

rprior

Description

Sample from the prior probability distribution.

Usage

## S4 method for signature 'pomp'
rprior(object, params, ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
additional arguments are ignored.

Value

A numeric matrix containing the required samples.

See Also

Specification of the prior distribution simulator: prior_spec

Other pomp workhorses: dmeasure(), dprior(), dprocess(), flow(), partrans(), rinit(), rmeasure(), rprocess(), skeleton(), workhorses

[Package pomp version 3.1.0.1 Index]