rprocess {pomp}	R Documentation

rprocess

Description

rprocess simulates the process-model portion of partially-observed Markov process.

Usage

## S4 method for signature 'pomp'
rprocess(object, x0, t0, times, params, ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

x0
an nvar x nrep matrix containing the starting state of the system. Columns of x0 correspond to states; rows to components of the state vector. One independent simulation will be performed for each column. Note that in this case, params must also have nrep columns.

t0
the initial time, i.e., the time corresponding to the state in x0.

times
a numeric vector (length ntimes) containing times. These must be in non-decreasing order.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x0.

...
additional arguments are ignored.

Details

When rprocess is called, t0 is taken to be the initial time (i.e., that corresponding to x0). The values in times are the times at which the state of the simulated processes are required.

Value

rprocess returns a rank-3 array with rownames. Suppose x is the array returned. Then

dim(x)=c(nvars,nrep,ntimes),

where nvars is the number of state variables (=nrow(x0)), nrep is the number of independent realizations simulated (=ncol(x0)), and ntimes is the length of the vector times. x[,j,k] is the value of the state process in the j-th realization at time times[k]. The rownames of x will correspond to those of x0.

See Also

Specification of the process-model simulator: rprocess_spec

More on pomp workhorse functions: dmeasure(), dprior(), dprocess(), flow(), partrans(), rinit(), rmeasure(), rprior(), skeleton(), workhorses

[Package pomp version 3.5.0.0 Index]