saved.states {pomp}	R Documentation

Saved states

Description

Retrieve latent state trajectories from a particle filter calculation.

Usage

## S4 method for signature 'pfilterd_pomp'
saved.states(object, ...)

## S4 method for signature 'pfilterList'
saved.states(object, ...)

Arguments

object
result of a filtering computation

...
ignored

Details

When one calls pfilter with save.states=TRUE, the latent state vector associated with each particle is saved. This can be extracted by calling saved.states on the ‘pfilterd.pomp’ object.

Value

The saved states are returned in the form of a list, with one element per time-point. Each element consists of a matrix, with one row for each state variable and one column for each particle.

See Also

More on particle-filter based methods in pomp: bsmc2(), cond.logLik(), eff.sample.size(), filter.mean(), filter.traj(), kalman, mif2(), pfilter(), pmcmc(), pred.mean(), pred.var(), wpfilter()

[Package pomp version 3.1.1.12 Index]