skeleton {pomp}	R Documentation

skeleton

Description

Evaluates the deterministic skeleton at a point or points in state space, given parameters. In the case of a discrete-time system, the skeleton is a map. In the case of a continuous-time system, the skeleton is a vectorfield. NB: skeleton just evaluates the deterministic skeleton; it does not iterate or integrate (see trajectory for this).

Usage

## S4 method for signature 'pomp'
skeleton(object, x, times, params, ...)

Arguments

object
an object of class ‘pomp’, or of a class that extends ‘pomp’. This will typically be the output of pomp, simulate, or one of the pomp inference algorithms.

x
an array containing states of the unobserved process. The dimensions of x are nvars x nrep x ntimes, where nvars is the number of state variables, nrep is the number of replicates, and ntimes is the length of times. One can also pass x as a named numeric vector, which is equivalent to the nrep=1, ntimes=1 case.

times
a numeric vector (length ntimes) containing times. These must be in non-decreasing order.

params
a npar x nrep matrix of parameters. Each column is treated as an independent parameter set, in correspondence with the corresponding column of x.

...
additional arguments are ignored.

Value

skeleton returns an array of dimensions nvar x nrep x ntimes. If f is the returned matrix, f[i,j,k] is the i-th component of the deterministic skeleton at time times[k] given the state x[,j,k] and parameters params[,j].

See Also

Specification of the deterministic skeleton: skeleton_spec

Other pomp workhorses: dmeasure(), dprior(), dprocess(), flow(), partrans(), rinit(), rmeasure(), rprior(), rprocess(), workhorses

[Package pomp version 3.1.1.6 Index]