time {pomp}	R Documentation

Methods to manipulate the obseration times

Description

Get and set the vector of observation times.

Usage

## S4 method for signature 'pomp'
time(x, t0 = FALSE, ...)

## S4 replacement method for signature 'pomp'
time(object, t0 = FALSE, ...) <- value

Arguments

x
a ‘pomp’ object

t0
logical; should the zero time be included?

...
ignored

object
a ‘pomp’ object

value
numeric vector; the new vector of times

Details

time(object) returns the vector of observation times. time(object,t0=TRUE) returns the vector of observation times with the zero-time t0 prepended.

time(object) <- value replaces the observation times slot (times) of object with value. time(object,t0=TRUE) <- value has the same effect, but the first element in value is taken to be the initial time. The second and subsequent elements of value are taken to be the observation times. Those data and states (if they exist) corresponding to the new times are retained.

[Package pomp version 3.1.1.1 Index]