traces {pomp}	R Documentation

Traces

Description

Retrieve the history of an iterative calculation.

Usage

## S4 method for signature 'mif2d_pomp'
traces(object, pars, transform = FALSE, ...)

## S4 method for signature 'mif2List'
traces(object, pars, ...)

## S4 method for signature 'abcd_pomp'
traces(object, pars, ...)

## S4 method for signature 'abcList'
traces(object, pars, ...)

## S4 method for signature 'pmcmcd_pomp'
traces(object, pars, ...)

## S4 method for signature 'pmcmcList'
traces(object, pars, ...)

Arguments

object
an object of class extending ‘pomp’, the result of the application of a parameter estimation algorithm

pars
names of parameters

transform
logical; should the traces be transformed back onto the natural scale?

...
ignored or passed to the more primitive function

Details

Note that pmcmc does not currently support parameter transformations.

Value

When object is the result of a mif2 calculation, traces(object, pars) returns the traces of the parameters named in pars. By default, the traces of all parameters are returned. If transform=TRUE, the parameters are transformed from the natural scale to the estimation scale.

When object is a ‘abcd_pomp’, traces(object) extracts the traces as a coda::mcmc.

When object is a ‘abcList’, traces(object) extracts the traces as a coda::mcmc.list.

When object is a ‘pmcmcd_pomp’, traces(object) extracts the traces as a coda::mcmc.

When object is a ‘pmcmcList’, traces(object) extracts the traces as a coda::mcmc.list.

See Also

Other extraction methods: coef(), cond.logLik(), covmat(), eff.sample.size(), filter.mean(), filter.traj(), forecast(), logLik, obs(), pred.mean(), pred.var(), saved.states(), spy(), states(), summary(), timezero(), time()

[Package pomp version 4.0.11.1 Index]