traces {pomp}	R Documentation

Traces

Description

Retrieve the history of an iterative calculation.

Usage

## S4 method for signature 'mif2d_pomp'
traces(object, pars, transform = FALSE, ...)

## S4 method for signature 'mif2List'
traces(object, pars, ...)

## S4 method for signature 'abcd_pomp'
traces(object, pars, ...)

## S4 method for signature 'abcList'
traces(object, pars, ...)

## S4 method for signature 'pmcmcd_pomp'
traces(object, pars, ...)

## S4 method for signature 'pmcmcList'
traces(object, pars, ...)

Arguments

object
an object of class extending ‘pomp’, the result of the application of a parameter estimation algorithm

pars
names of parameters

transform
logical; should the traces be transformed back onto the natural scale?

...
ignored or (in the case of the listie, passed to the more primitive function)

Details

Note that pmcmc does not currently support parameter transformations.

Value

When object is the result of a mif2 calculation, traces(object, pars, transform = FALSE) returns the traces of the parameters named in pars. By default, the traces of all parameters are returned. Note that, if the computation was performed with transformed parameters, the traces are on the estimation scale. If transform=TRUE, the parameters are transformed from the estimation scale onto the natural scale.

When object is a ‘abcd_pomp’, traces(object) extracts the traces as a coda::mcmc.

When object is a ‘abcList’, traces(object) extracts the traces as a coda::mcmc.list.

When object is a ‘pmcmcd_pomp’, traces(object) extracts the traces as a coda::mcmc.

When object is a ‘pmcmcList’, traces(object) extracts the traces as a coda::mcmc.list.

[Package pomp version 2.8.0.3 Index]