verhulst {pomp}	R Documentation

Verhulst-Pearl model

Description

The Verhulst-Pearl (logistic) model of population growth.

Usage

verhulst(n_0 = 10000, K = 10000, r = 0.9, sigma = 0.4, tau = 0.1, dt = 0.01)

Arguments

n_0
initial condition

K
carrying capacity

r
intrinsic growth rate

sigma
environmental process noise s.d.

tau
measurement error s.d.

dt
Euler time-step

Details

A stochastic version of the Verhulst-Pearl logistic model. This evolves in continuous time, according to the stochastic differential equation

dn = r n (1-n/K) dt + sigma n dW.

Numerically, we simulate the stochastic dynamics using an Euler approximation.

The measurements are assumed to be log-normally distributed.

Value

A ‘pomp’ object containing the model and simulated data. The following basic components are included in the ‘pomp’ object: ‘rinit’, ‘rprocess’, ‘rmeasure’, ‘dmeasure’, and ‘skeleton’.

See Also

More examples provided with pomp: blowflies, bsflu, dacca(), ebola, gompertz(), measles, ou2(), parus, pomp_examples, ricker(), rw2(), sir_models

Examples

verhulst() -> po
plot(po)
plot(simulate(po))
pfilter(po,Np=1000) -> pf
logLik(pf)
spy(po)

[Package pomp version 3.2.0.0 Index]