workhorses {pomp}	R Documentation

Workhorse functions for the pomp algorithms.

Description

These functions mediate the interface between the user's model and the package algorithms. They are low-level functions that do the work needed by the package's inference methods.

Details

They include

dmeasure

which evaluates the measurement model density,

rmeasure

which samples from the measurement model distribution,

dprocess

which evaluates the process model density,

rprocess

which samples from the process model distribution,

dprior

which evaluates the prior probability density,

rprior

which samples from the prior distribution,

skeleton

which evaluates the model's deterministic skeleton,

flow

which iterates or integrates the deterministic skeleton to yield trajectories,

partrans

which performs parameter transformations associated with the model.

Author(s)

Aaron A. King

See Also

basic model components, elementary algorithms, estimation algorithms

More on pomp workhorse functions: dmeasure(), dprior(), dprocess(), flow(), partrans(), rinit(), rmeasure(), rprior(), rprocess(), skeleton()

[Package pomp version 3.4 Index]